requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
dnspython==2.3.0
